+++ src/settings/constants.ts
@@ -1,0 +1,8 @@
+export enum SETTINGS {
+	OptimizeTokenUiBatching = 'optimize-interface-layer-clipping',
+	EffectsCaching = 'token-effects-caching',
+	TokenBarsCaching = 'token-bars-caching',
+	SpritesheetSubstitution = 'spritesheet-substitution',
+	TokenRingSpritesheetSupport = 'token-ring-spritesheet-support',
+	CustomSpritesheets = 'spritesheet-substitution-custom',
+}